requirements.txt
discord.py==2.6.4
aiohttp==3.13.2
beautifulsoup4==4.14.3
python-dotenv==1.0.0
flask==3.0.0
psycopg2-binary==2.9.9

